+++ src/com/dci/intellij/dbn/object/DBTypeAttribute.java
@@ -6,5 +6,5 @@ import com.dci.intellij.dbn.object.common.DBObject;
 public interface DBTypeAttribute extends DBObject {
     DBType getType();
     DBDataType getDataType();
-    int getPosition();
+    short getPosition();
 }
